update: 開催報告一覧ページを (list) でグルーピング
